requirements.txt
opencv-python
numpy
tqdm
Pillow
imagehash
scikit-learn
